Move internal functions from connect.ts to utils
